done with all RQs
I've changed the code such that all python code is executed within a Rmd file. This means that all output that would go in the console is saved in the html output of the Rmd. This way, we have the advantage of keeping all output, but remove the disadvantages associated with Jupyter notebooks.